requirements.txt
azure-functions
yfinance
pandas
pyodbc
azure-storage-blob
python-dotenv
